WIP: Moving over to the STM32

+++ JuicePDM/src/ChannelConfig.h
@@ -28,10 +28,10 @@
 /// @brief Defines available channel types
 enum ChannelType
 {
-  DIG,              // Digital input
-  DIG_PWM,          // Digital input, PWM output
-  CAN_DIGITAL,      // CAN bus controlled digital output
-  CAN_PWM           // CAN bus controlled PWM output
+  DIG,         // Digital input
+  DIG_PWM,     // Digital input, PWM output
+  CAN_DIGITAL, // CAN bus controlled digital output
+  CAN_PWM      // CAN bus controlled PWM output
 };
 
 /// @brief Channel config structure
@@ -50,7 +50,7 @@ struct __attribute__((packed)) ChannelConfig
   float RetryDelay;           // Retry delay in seconds
   uint8_t MultiChannel;       // Grouped with other channels. Allows higher current loads
   uint8_t GroupNumber;        // Group membership number
-  uint8_t ControlPin;         // Digital uC control pin
+  int ControlPin;             // Digital uC control pin
   uint8_t CurrentSensePin;    // Current sense input pin
   uint8_t InputControlPin;    // Digital input control pin (digital channels only)
   uint8_t ErrorFlags;         // Bitmask for channel error flags
